# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Admin API for the AlefGym loyalty-club dashboard: customer segmentation, complaint verification, renewal tracking, and AI-assisted analysis.",
        "title": "Centropy Affiliate Admin API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/renewals/overdue": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Customers whose last delivered program is older than the given days, with zero completed orders since.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "renewals"
                ],
                "summary": "List overdue renewals",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "minimum days overdue (default 50)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/centropy-affilate_internal_domain_renewal.OverdueCustomer"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/centropy-affilate_internal_interfaces_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/segments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Six-segment breakdown (newcomer/cold/hero/at-risk/churned/one-time) plus total customer count.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "Segment summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/centropy-affilate_internal_domain_segment.Summary"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/centropy-affilate_internal_interfaces_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/segments/non-purchasers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Every registered customer who has never completed a single order.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "List non-purchasers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/centropy-affilate_internal_domain_segment.NonPurchaser"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/centropy-affilate_internal_interfaces_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/segments/non-purchasers/monthly": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Per signup month, how many of that month's registrants still have zero completed orders today.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "Monthly non-purchaser signups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/centropy-affilate_internal_domain_segment.MonthlySignups"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/centropy-affilate_internal_interfaces_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/segments/{segment}/customers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Every customer currently in one segment.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "List customers in a segment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "segment name (newcomer/cold/hero/at-risk/churned/one-time)",
                        "name": "segment",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/centropy-affilate_internal_domain_segment.Customer"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/centropy-affilate_internal_interfaces_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "centropy-affilate_internal_domain_renewal.OverdueCustomer": {
            "type": "object",
            "properties": {
                "completedOrdersTotal": {
                    "type": "integer"
                },
                "daysSinceLastProgram": {
                    "type": "integer"
                },
                "externalUserID": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "lastProgramAt": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "centropy-affilate_internal_domain_segment.Count": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "segment": {
                    "$ref": "#/definitions/centropy-affilate_internal_domain_segment.Segment"
                }
            }
        },
        "centropy-affilate_internal_domain_segment.Customer": {
            "type": "object",
            "properties": {
                "completedOrders": {
                    "type": "integer"
                },
                "externalUserID": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "centropy-affilate_internal_domain_segment.MonthlySignups": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "month": {
                    "type": "string"
                }
            }
        },
        "centropy-affilate_internal_domain_segment.NonPurchaser": {
            "type": "object",
            "properties": {
                "externalUserID": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "registeredAt": {
                    "type": "string"
                }
            }
        },
        "centropy-affilate_internal_domain_segment.Segment": {
            "type": "string",
            "enum": [
                "NEWCOMER",
                "COLD",
                "HERO",
                "AT_RISK",
                "CHURNED",
                "ONE_TIME"
            ],
            "x-enum-varnames": [
                "Newcomer",
                "Cold",
                "Hero",
                "AtRisk",
                "Churned",
                "OneTime"
            ]
        },
        "centropy-affilate_internal_domain_segment.Summary": {
            "type": "object",
            "properties": {
                "counts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/centropy-affilate_internal_domain_segment.Count"
                    }
                },
                "totalCustomers": {
                    "type": "integer"
                }
            }
        },
        "centropy-affilate_internal_interfaces_http_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT from /auth/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}